made edits after suggestions from evaluators

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: ka-tan <ka-tan@student.42singapore.sg>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/16 13:26:54 by ka-tan            #+#    #+#             */
-/*   Updated: 2025/05/19 19:24:44 by ka-tan           ###   ########.fr       */
+/*   Updated: 2025/06/15 17:04:20 by ka-tan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_atoi(const char *str)
 {
-	long	i;
+	size_t	i;
 	long	result;
 	long	sign;
 
@@ -42,19 +42,19 @@ int	ft_atoi(const char *str)
 // 	char	str2[] = "-456";
 // 	char	str3[] = "abc123";
 // 	char	str4[] = "+789";
-//     char	str5[] = "+++--8987";
-//     char	str6[] = "4547545+-+-";
+// 	char	str5[] = "+++--8987";
+// 	char	str6[] = "4547545+-+-";
 
 // 	printf("      123: %d\n", ft_atoi(str1));
-//     printf("      123: %d\n", atoi(str1));
+// 	printf("      123: %d\n", atoi(str1));
 // 	printf("-456: %d\n", ft_atoi(str2));
-//     printf("-456: %d\n", atoi(str2));
+// 	printf("-456: %d\n", atoi(str2));
 // 	printf("abc123: %d\n", ft_atoi(str3));
-//     printf("abc123: %d\n", atoi(str3));
+// 	printf("abc123: %d\n", atoi(str3));
 // 	printf("+789: %d\n", ft_atoi(str4));
 // 	printf("+789: %d\n", atoi(str4));
-//     printf("+++--8987: %d\n", ft_atoi(str5));
-//     printf("+++--8987: %d\n", atoi(str5));
+// 	printf("+++--8987: %d\n", ft_atoi(str5));
+// 	printf("+++--8987: %d\n", atoi(str5));
 // 	printf("4547545+-+-: %d\n", ft_atoi(str6));
-//     printf("4547545+-+-: %d\n", atoi(str6));
+// 	printf("4547545+-+-: %d\n", atoi(str6));
 // }
